equivalent domains APIs and data models

+++ src/Core/Enums/CipherType.cs
@@ -1,6 +1,6 @@
 ﻿namespace Bit.Core.Enums
 {
-    public enum CipherType : short
+    public enum CipherType : byte
     {
         Folder = 0,
         Login = 1
