+++ src/Core/Enums/DeviceType.cs
@@ -1,6 +1,6 @@
 ﻿namespace Bit.Core.Enums
 {
-    public enum DeviceType : short
+    public enum DeviceType : byte
     {
         Android = 0,
         iOS = 1
